requirements.txt
attrs>=20.3.0
cycler>=0.10.0
iniconfig>=1.1.1
kiwisolver>=1.3.1
matplotlib>=3.1.3
nose>=1.3.7
numpy>=1.19.5
openpyxl>=3.0.7
packaging>=20.8
Pillow>=8.2.0
pluggy>=0.13.1
py>=1.10.0
pyparsing>=2.4.7
pytest>=6.2.1
python-dateutil>=2.8.1
pytz>=2020.5
scipy>=1.6.0
seaborn>=0.11.1
six>=1.15.0
toml>=0.10.2
typing-extensions>=3.7.4.3
xlsxwriter>=1.4.3
sphinx-rtd-theme>=0.4
myst-nb>=0.13.1
autograd>=1.3
jax>=0.3.4
jaxlib>=0.3.4
equinox>=0.9
numpyro>=0.1
arviz>=0.13
optax>=0.1
multimethod>=1.11
PyYAML~=6.0.2
inferactively-pymdp